add exchangeRates api with quartz job

+++ src/main/java/hr/tvz/financije/controllers/models/commands/AccountCommand.java
@@ -8,7 +8,7 @@ import jakarta.validation.constraints.Size;
 public record AccountCommand(
         @Positive Integer id,
         @NotBlank @Size(max = 50) String name,
-        @NotNull @Positive Integer amount,
+        @NotNull @Positive Long amount,
         @NotBlank @Size(max = 10) String type,
         @NotBlank @Size(max = 7) String color,
         @NotBlank @Size(max = 10) String source,
